requirements.txt
pandas

numpy

scikit-learn

streamlit

mysql-connector-python

psycopg2-binary


pymysql

sqlalchemy


beautifulsoup4

